Bundle size optimisation
- ts configuration target changed from `es5` to `es2017`
- moved Default SEO configuration from `_app.tsx` to `@components/layout`
- highlight.js configuration is now dynamically loaded
- Canonical URL for dynamic routes fixed
