+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "next-notion-portfolio",
-    "version": "1.1.0",
+    "version": "1.1.1",
     "private": true,
     "scripts": {
         "dev": "next dev",
