+++ src/components/layout/index.tsx
@@ -1,6 +1,10 @@
 import { Footer } from "@components/footer";
 
+import { getBaseURL } from "@util/router";
+
 import { motion } from "framer-motion";
+import { DefaultSeo } from "next-seo";
+import { useRouter } from "next/router";
 import { ReactNode } from "react";
 
 type Props = {
@@ -14,8 +18,32 @@ const variants = {
 };
 
 export const Layout: React.FC<Props> = ({ children }: Props) => {
+    const router = useRouter();
+    const baseUrl = getBaseURL();
+    const url = `${baseUrl}${router.asPath}`;
     return (
         <div>
+            <DefaultSeo
+                titleTemplate="%s - Alan John"
+                openGraph={{
+                    type: "website",
+                    locale: "en_IE",
+                    url: url,
+                    description:
+                        "The personal website for Alan John, developer.",
+                    site_name: "Alan John",
+                    images: [
+                        {
+                            url: `${baseUrl}/images/social_media_preview.png`,
+                            width: 1200,
+                            height: 628,
+                            alt: "My Portfolio Preview",
+                            type: "image/png",
+                        },
+                    ],
+                }}
+                canonical={url}
+            />
             <motion.main
                 initial="hidden"
                 animate="enter"
